Change components to ui and separate state from iced logic
