--- crates/gui/src/debugger/components.rs
@@ -1,3 +1,0 @@
-pub mod register;
-
-pub use register::Registers;
