+++ crates/gui/src/debugger/ui/register.rs
